Reorganization scripts files
