+++ scripts/push_sdk.py
@@ -1,11 +1,11 @@
 import sys, getopt
 
-from .shared.clients.fs_client import FsClient
-from .shared.clients.git_client import GitClient
-from .shared.clients.os_client import OsClient
-from .languages.php.php_sdk_push_action import PhpSdkPushAction
-from .languages.default.default_push_action import DefaultPushSdkAction
-from .shared.utils import get_logger
+from shared.clients.fs_client import FsClient
+from shared.clients.git_client import GitClient
+from shared.clients.os_client import OsClient
+from languages.php.php_sdk_push_action import PhpSdkPushAction
+from languages.default.default_push_action import DefaultPushSdkAction
+from shared.utils import get_logger
 
 logger = get_logger()
