+++ scripts/shared/clients/os_client.py
@@ -2,7 +2,7 @@ from os import path
 
 from ..models.programming_language import ProgrammingLanguage
 from ..models.criteo_service import CriteoService
-from ...shared.utils import assert_environment_variable
+from ..utils import assert_environment_variable
 
 
 class IOsClient:
